modularizacion y cambios menores
