+++ appanimals/lib/widgets/animal_history.dart
@@ -18,9 +18,7 @@ class AnimalHistory extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Center(
-          child: Text('Historia del Animal')
-        ),
+        title: Text('Historia del Animal'),
         backgroundColor: const Color.fromARGB(255, 21, 100, 21),
         foregroundColor: Colors.white,
       ),
